add status table for class_refinement and class_main_dft it will spit out ASCII table using tabulate package
